Bouge updateDPC dans un hook à part useUpdateDPC
